requirements.txt
fastapi
uvicorn
torch
ultralytics
Pillow
python-multipart